% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{current_resolver}
\alias{current_resolver}
\title{The resolver \code{read_source()} is currently dispatching to}
\usage{
current_resolver()
}
\value{
A function. Defaults to \code{\link[=oracle_reader]{oracle_reader()}}.
}
\description{
The resolver \code{read_source()} is currently dispatching to
}
