% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\docType{data}
\name{ORACLE_TABLES}
\alias{ORACLE_TABLES}
\title{Logical name to Oracle object}
\format{
A named character vector.
}
\usage{
ORACLE_TABLES
}
\description{
The SAS libnames \code{rpt} and \code{msow} correspond to these two schemas.
}
\keyword{datasets}
