% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_layout.R
\docType{data}
\name{TELEMED_IDENTITY_COLUMNS}
\alias{TELEMED_IDENTITY_COLUMNS}
\title{Practitioner-level columns of the weekly telemed report, in report order}
\format{
A character vector of 9 column names.
}
\usage{
TELEMED_IDENTITY_COLUMNS
}
\description{
These are also the crosstab index: one output row per distinct combination.
}
\seealso{
\link{TELEMED_LABELS}, \code{\link[=ordered_columns]{ordered_columns()}}
}
\keyword{datasets}
