% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle_utils.R
\name{the_connection}
\alias{the_connection}
\title{The configured connection}
\usage{
the_connection()
}
\value{
The connection passed to \code{\link[=configure]{configure()}}.
}
\description{
The configured connection
}
\seealso{
\code{\link[=configure]{configure()}}
}
